added like functionality

+++ src/pages/Mytips.jsx
@@ -1,10 +1,9 @@
 import React, { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../context/AuthProvider";
 import { Link } from "react-router";
-import { FaEye, FaPen  } from "react-icons/fa";
+import { FaEye, FaPen } from "react-icons/fa";
 import { MdDelete } from "react-icons/md";
-
-
+import Swal from "sweetalert2";
 
 const Mytips = () => {
   const { user } = useContext(AuthContext);
@@ -19,24 +18,45 @@ const Mytips = () => {
       });
   }, [user.email]);
 
-    const handleDelete=(_id)=>{
-    fetch(`http://localhost:3000/delete/${_id}`,{
-      method:"DELETE"
-    }).then(res=>res.json())
-    .then(data=>{console.log(data)})
-  }
+  const handleDelete = (_id) => {
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      console.log(result.isConfirmed);
+      if (result.isConfirmed) {
+        fetch(`http://localhost:3000/delete/${_id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then(() => {
+            const remainingtip = all.filter((item) => item._id !== _id);
+            setAll(remainingtip);
+          });
+      }
+    });
+  };
   return (
     <div>
-      <div className="overflow-x-auto rounded-md border border-black">
+      <div
+        className="overflow-x-auto rounded-md border border-black container mx-auto lg:min-h-[calc(100vh-500px)] my-10">
         <table className="table w-full">
           <thead>
             <tr className="border-b border-black ">
               <th>Title</th>
               <th>category</th>
-              <th></th>
+              <th className="hidden md:block">See More</th>
+              <th>Update</th>
+              <th>Delete</th>
             </tr>
           </thead>
           <tbody>
+            
             {all.map((item) => (
               <tr key={item.id} className="border-t border-black ">
                 <td>
@@ -52,17 +72,24 @@ const Mytips = () => {
                   </div>
                 </td>
                 <td>{item.category}</td>
-                <td>
+                <td className="hidden md:block">
                   <Link to={`/tipdetails/${item._id}`}>
                     <FaEye className="btn btn-ghost border-none size-14" />
                   </Link>
+                </td>
+                <td>
+                  {" "}
                   <Link to={`/updatemytip/${item._id}`}>
                     <FaPen className="btn btn-ghost border-none size-14" />
                   </Link>
+                </td>
+                <td>
                   <div>
-                    <MdDelete  onClick={()=>handleDelete(item._id)} className="btn btn-ghost border-none size-14" />
+                    <MdDelete
+                      onClick={() => handleDelete(item._id)}
+                      className="btn btn-ghost border-none size-14"
+                    />
                   </div>
-                
                 </td>
               </tr>
             ))}
